wordle: Add `dict::default()` method for convenience
Refactor SliceDict and HashDict into their own modules.
Expose a `default()` method in `dict` that returns a static reference to
a SliceDict with the default word lists.
